Update README.md
A system that simulates a library. The user can add a book to the system and specify the name of the book and the name of the author. He can also borrow and return it. The system allows the user to know the number of books available, and to know if the book is borrowed or not.